todolist/requirements.txt
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.4
Markdown
django-environ==0.4.5
pytz==2021.1
